Different Arrow Syntaxes
So the arrow syntax is nothing Earth-shattering. But it's used a lot, so you need to train your eyes to recognize that it's just an anonymous function.

And sometimes, it can look a bit different. For example, the parentheses around the arguments? Totally optional! Without them, everything still works.
I like the parentheses: I feel like it gives my arrow functions a bit more structure. But other code might not have them.