Arrow Functions Everywhere!
Let's use the arrow syntax everywhere! Below, when we use SweetAlert, remove the self variable and - in preConfirm - use () for empty arguments, then add the arrow. Inside, we can use this! Use the arrow function again below. Here, we're not using this, but I like to stay consistent and use the arrow function everywhere.