Keep going! Inside the next method, remove self, and add our arrow function.
Do the same for fadeOut(). But here, we were using this, which previously pointed to the DOM Element object that was fading out. We can't use this anymore, but that's fine! Replace it with {dollar_sign}row.remove() and then this.updateTotalWeight().
Double-check that things work. Refresh! Delete one of the items and... perfect!