The ECMAScript gods didn't stop with let - they added a third way to declare a new variable: var, let and const.
Back in our play file, remove the log and initialize the variable with const.

As you're probably already guessing - and as you can see from PhpStorm being very angry - when you initialize a variable with const, it can't be reassigned.

When we try this now, we get a huge error!

TypeError: Assignment to constant variable.