var, let and const
Phew! Okay, in ES2015, var, let, and const do have some differences. But a lot of the time, you will see people use one or the other purely based on personal preference, or whatever is trendy that week. The differences are minor.

In our case, because you know I love to write hipster code, let's change each let to const. Start at the top: const {dollar_sign}link makes sense. We don't need to reassign that. The same is true for deleteUrl, {dollar_sign}row, {dollar_sign}form, and formData.

Sure, we modify formData, but we know that's ok with const!