Let's see one more example: scroll all the way down to the Helper class. Create a new method: static _calculateWeight() with an {dollar_sign}elements argument. This will be a new static utility method whose job is to loop over whatever elements I pass, look for their weight data attribute, and then return the total weight. We don't really need to make this change, but it's valid.
Now, in calculateTotalWeight(), just say: return Helper - because we need to reference the static method by its class name Helper._calculateTotalWeight() and pass it the elements: this.{dollar_sign}wrapper.find('tbody tr').
Coolio! Try that out! And we still see the correct total.

And that is 10 times easier to understand as a PHP developer! Sure, JavaScript still has prototypical inheritance behind the scenes... but most of the time, we won't know or care.