Destructuring & Default Values
What if we try to create a variable that does not match a key, like, I don't know, totallyMadeUpKey.

Try to print that. What do you think will happen?

It's not an error: it just prints as undefined. So destructuring is friendly: if something goes wrong, it doesn't kill your code: it just assigns undefined.