Let's look for the other {dollar_sign}.each() spots and update those too! Instead, say for let fieldData of {dollar_sign}form.serializeArray(). Before, the anonymous function received a key and then the fieldData. But, we didn't actually need the key: the {dollar_sign}.each() function just forced us to add it. Now, things are cleaner!
Make this same change in two more places: for {dollar_sign}element of {dollar_sign}form.find(':input'). Ah, don't forget your let or var.
Then, one more below: for let  of {dollar_sign}elements.

So, use the for of loop for everything! Well actually, that's not 100% true. for of is perfect when you want to loop over a collection of items. But, if you want to loop over an associative array... or object, and you need to know the key for each item, then you'll use for in.

 Tip
Actually, you can use for of with an object, with a clever combination of Object.entries() and array destructuring!

let pets = {
  beagle: 'Bark Twain',
  poodle: 'Snuffles'
};

for (let [petKey, petName] of Object.entries(pets)) {
  console.log(petKey, petName);
}
BUT, the Object.entries() method is still experimental, and may be included in ES2017.

This is the one limitation of for of: it gives you the value of the item you're looping over, but not its key, or index. In fact, if try to use for of with an object, you'll get an error.