Map with Non-String Keys
Map has one other advantage... which is kind of crazy: you can use non-string keys!

Try this: create a new variable: let southernUSStates set to an array of Tennessee, Kentucky, and Texas.
Now we can say foods.set(southernUSStates) and set that to hot chicken.
Yes, the key is actually an object. And that's no problem!
Anyways, at the bottom, use foods.get(southernUSStates) to fetch out that value.
And it works just like we want!