WeakMap has an extra requirement: its keys must be objects.
Turn each key into an array, which is an object. At the bottom, use foods.get() and pass it the italian array.

Now when I run it, it works fine. Wait, or, does it?

Two interesting things: this prints undefined, hot chicken, undefined. First, even though the ['italian'] array in get() is equal to the ['italian'] array used in set, they are not the same object in memory. These are two distinct objects, so it looks like a different key to WeakMap. That's why it prints undefined.

Second, with WeakMap, you can't call foods.size. That's just not something that works with WeakMap.