To see a real-world WeakMap use-case, go back into RepLogApp and scroll to the top. Remember, this file holds two classes: RepLogApp and, at the bottom, Helper: 233 lines  web/assets/js/RepLogApp.js ... lines 1 - 2 (function(window, {dollar_sign}, Routing, swal) {   class RepLogApp { ... lines 6 - 178   }   /**    * A 'private' object    */     class Helper { ... lines 185 - 212     } ... lines 214 - 231 })(window, jQuery, Routing, swal);
The purpose of Helper is to be a private object that we can only reference from inside of this self-executing function.

Making Helper a Private Object

But check out the constructor for RepLogApp:

 233 lines  web/assets/js/RepLogApp.js
... lines 1 - 2
(function(window, {dollar_sign}, Routing, swal) {
  class RepLogApp {
    constructor({dollar_sign}wrapper) {
... line 7
      this.helper = new Helper(this.{dollar_sign}wrapper);
... lines 9 - 26
    }
... lines 28 - 178
  }
... lines 180 - 231
})(window, jQuery, Routing, swal);
We set Helper onto a helper property. We do this so that we can use it later, inside of updateTotalWeightLifted(). Here's the problem: the helper property is not private. I mean, inside of our template, if we wanted, we could say: repLogApp.helper.calculateTotalWeight().

Dang! We went to all of that trouble to create a private Helper object... and it's not actually private! Lame!