But now, back in RepLogApp, after we set the HelperInstance, simply return.
This is a temporary hack to show off garbage collection. Now that we're returning immediately, when we create a new RepLogApp object, it's not attaching any listeners or adding itself as a reference to anything in the code. In other words, this object is not attached or referenced anywhere in memory. Because of that, RepLogApp objects - and their Helper objects - should be eligible for garbage collection.

Now, garbage collection isn't an instant process - it takes places at intervals, and it's up to your JavaScript engine to worry about that. But if you're using Chrome, you can force garbage collection! On the timeline tab, you should see a little garbage icon. Try this: refresh! Quickly click the 'collect garbage' button, and then see what prints in the console.

Ok, so HelperInstances still has 5 objects inside. In other words, the Helper objects were not garbage collected. Why? Because they are still being referenced in the code... by the Map itself!