Now, change the Map to a WeakMap. Go back and repeat the dance: refresh, hit the garbage icon, and then go to the console. Woh! Check this out! The WeakMap is empty. Remember, this is its superpower! Since none of the RepLogApp objects are being referenced in memory anymore, both those and their Helper instances are eligible for garbage collection. When you use Map, it prevents this: simply being inside of the Map counts as a reference. With WeakMap that doesn't happen.
Ok, I know, this was still pretty darn advanced. So you may or may not have this use case. But this is when you will see WeakMap used instead of Map. For us it means we should use Map in normal situations... and WeakMap only if we find ourselves with this problem.