The Map object is perfect for maps, or associative arrays as we call them in the PHP biz. But what about true, indexed arrays? Well actually, JavaScript has always had a great way to handle these - it's not new! It's the Array object.
Well, the Array object isn't new, but it does have a new trick. Let's check out an example: when the page loads, we call loadRepLogs():
235 lines  web/assets/js/RepLogApp.js
... lines 1 - 2
(function(window, {dollar_sign}, Routing, swal) {
... lines 4 - 6
  class RepLogApp {
    constructor({dollar_sign}wrapper) {
... lines 9 - 11
    this.loadRepLogs();
... lines 13 - 28
   }
... lines 30 - 39
   loadRepLogs() {
     {dollar_sign}.ajax({
       url: Routing.generate('rep_log_list'),
     }).then(data => {
       for (let repLog of data.items) {
         this._addRow(repLog);
       }
     })
   }
... lines 49 - 170
   _addRow(repLog) {
... lines 172 - 175
     const html = rowTemplate(repLog);
     this..find('tbody').append({dollar_sign}.parseHTML(html));
     this.updateTotalWeightLifted();
   }
 }
... lines 182 - 233
})(window, jQuery, Routing, swal);
This fetches an array of repLog data via AJAX and then calls _addRow() on each to add the <tr> elements to the table.

But once we add the table rows... we don't actually store those repLog objects anywhere. Yep, we use them to build the page... then say: Adios!

Now, I do want to start storing this data on my object, and you'll see why in a minute